--- UnifiedLogger/UnifiedLogger.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net8.0-ios; net8.0-maccatalyst</TargetFrameworks>
-    <SingleProject>true</SingleProject>
-    <ImplicitUsings>disable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <PackageId>Mahonkin.UnifiedLogger</PackageId>
-    <Description>Implementation of ILoggerProvider for logging to Apple Unified Logging on Apple devices.</Description>
-    <Version>0.0.11</Version>
-    <Authors>Timothy Reynolds</Authors>
-    <Company></Company>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-    <PackageOutputPath>..\..\nupkg</PackageOutputPath>
-    <PackageTags>maui;macos;ios;oslog;logging;log</PackageTags>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Include="README.md" Pack="true" PackagePath="\" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Configuration" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Maui.Controls" Version="8.*" />
-  </ItemGroup>
-</Project>
