+++ UnifiedLogging.sln
@@ -2,7 +2,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnifiedLogger", "UnifiedLogger\UnifiedLogger.csproj", "{D31CCF0E-20E7-421B-8398-53207117CD86}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnifiedLogging", "UnifiedLogging\UnifiedLogging.csproj", "{D31CCF0E-20E7-421B-8398-53207117CD86}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
