+++ UnifiedLogging/TestProgram.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace com.mahonkin.tim.UnifiedLogging;
+
+public class TestProgram
+{
+    public static void Main(string[] args)
+    {
+    }
+}
