+++ UnifiedLogging/UnifiedLogging.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <OutputType>Exe</OutputType>
+    <ImplicitUsings>disable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <!-- <PackageId>Mahonkin.UnifiedLogger</PackageId>
+    <Description>Implementation of ILoggerProvider for Apple Unified Logging.</Description>
+    <Version>0.0.11</Version>
+    <Authors>Timothy Reynolds</Authors>
+    <Company></Company>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <PackageOutputPath>..\..\nupkg</PackageOutputPath>
+    <PackageTags>macos;ios;oslog;logging;log</PackageTags> -->
+  </PropertyGroup>
+  <ItemGroup>
+    <!-- <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" /> -->
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
+    <!-- <PackageReference Include="Microsoft.Extensions.Logging.Configuration" Version="8.0.0" /> -->
+  </ItemGroup>
+</Project>
