+++ UnifiedLogging/UnifiedLoggingConfig.cs
@@ -1,0 +1,5 @@
+namespace com.mahonkin.tim.UnifiedLogging;
+
+public static class UnifiedLoggingConfig
+{
+}
